% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dccav.r
\name{print.dccav}
\alias{print.dccav}
\title{Print a summary of a dc-CA object}
\usage{
\method{print}{dccav}(x, ...)
}
\arguments{
\item{x}{a dc-CA object from \code{\link{dc_CA_vegan}}}
}
\description{
Print a summary of a dc-CA object
}
\details{
\code{x <- print(x)} is more efficient for \code{\link{scores.dccav}} than just \code{print(x)}
if \code{\link{dc_CA_vegan}} is called without argument \code{verbose} (or called with \code{verbose = FALSE}).
}
