% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dccav.r
\name{print.dccav}
\alias{print.dccav}
\title{Print a summary of a dc-CA object}
\usage{
\method{print}{dccav}(x, ...)
}
\arguments{
\item{x}{a dc-CA object from \code{\link{dc_CA_vegan}}}

\item{...}{Other arguments passed to the function (currently ignored).}
}
\description{
Print a summary of a dc-CA object
}
\details{
\code{x <- print(x)} is more efficient for \code{\link{scores.dccav}} than just \code{print(x)}
if \code{\link{dc_CA_vegan}} is called without argument \code{verbose} (or called with \code{verbose = FALSE}).
}
\examples{

data("dune_trait_env")
# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites
# must delete "Sites" from response matrix or data frame
Y <- dune_trait_env$comm[,-1] # must delete "Sites"

out <- dc_CA_vegan(formulaEnv = ~A1+Moist+Use+Manure+Condition(Mag),
                   formulaTraits = ~ SLA + Height + LDMC + Seedmass +Lifespan,
                   response = Y,
                   dataEnv =dune_trait_env$envir,
                   dataTraits =dune_trait_env$traits,
                   verbose = FALSE)
plot_dcCA(out, verbose = FALSE)

}
