% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dccav.r
\name{print.dccav}
\alias{print.dccav}
\title{Print a summary of a dc-CA object}
\usage{
\method{print}{dccav}(x, ...)
}
\arguments{
\item{a}{dc-CA object from \code{\link{dc_CA_vegan}}}
}
\description{
Print a summary of a dc-CA object
}
