% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.dccav.R
\name{scores.dccav}
\alias{scores.dccav}
\title{Extract results of a vegan-based double constrained correspondence Analysis (dc-CA)}
\usage{
\method{scores}{dccav}(
  x,
  choices = c(1, 2),
  display = c("all"),
  which_cor = "in model",
  tidy = FALSE,
  ...
)
}
\arguments{
\item{x}{object}

\item{choices}{integer vector of which axes to obtain. Default: all dc-CA axes.}

\item{display}{a character vector, one or more of
\code{c("all","species","sites","sp", "wa", "lc","bp", "cor", "reg", "cn",
"lc_traits", "reg_traits", "cor_traits","bp_traits","cn_traits")}.
The first ten are as in \code{\link[vegan]{scores.cca}} (except \code{"cor"})
and remaining ones are similar scores for traits.}

\item{which_cor}{character or list of trait and environmental variables names in the data frames
for which inter-set correlations must calculated.
Default: a character ("in_model") for all traits and variables in the model,
including collinear variables and levels.}

\item{tidy}{Return scores that are compatible with \code{ggplot2}:
all scores are in a single data.frame, score type is identified by factor variable \code{score},
the names by variable \code{label}, and species weights (in dc_CA_vegan) are in variable \code{weight}.
See \code{\link[vegan]{scores.cca}}.}

\item{...}{Other arguments passed to the function (currently ignored).}
}
\description{
This function works very much like the \code{vegan} \code{\link[vegan]{scores}} function,
in particular \code{\link[vegan]{scores.cca}}, with the additional results such
as regression coefficients and linear combinations of traits \code{('regr_traits','lc_traits')}
In the current version, there is a single scaling (\code{scaling = "sites"}).
All scores from CA obey the so called transition formulas and so do the scores of CCA and dc-CA.
The differences are, for CCA, that the linear combinations of environmental variables
(the \emph{constrained} site scores)
replace the usual (\emph{unconstrained}) site scores, and for dc-CA,
that the linear combinations of traits (the \emph{constrained} species scores)
also replace the usual (\emph{unconstrained}) species scores.
}
\details{
In current version: no \code{biplot_traits} scores; \code{bp_traits} has no effect.
In current version: no \code{centroid_traits} scores;\code{cn_traits} has no effect.
}
\examples{

data("dune_trait_env")
# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites
mod <- dc_CA_vegan(formulaEnv = ~A1+Moist+Mag+Use+Manure,
                   formulaTraits = ~. ,
                   response = dune_trait_env$comm[, -1],  # must delete "Sites"
                   dataEnv =dune_trait_env$envir,
# delete "Species", "Species_abbr" from traits and
# use all remaining variables due to formulaTraits = ~. (the default)
                   dataTraits =dune_trait_env$traits[,-c(1,2)],
                   verbose = TRUE)


# community-level permutation test
anova(mod$RDAonEnv) # all option of anova.cca are available!
# a species-level permuation test requires an dedicated new function
# this is not yet available in this version

mod_scores <- scores(mod, display = "all")
names(mod_scores)

mod_scores <- scores(mod, display = "all", tidy = TRUE)
names(mod_scores)
levels(mod_scores$score)


# for illustration: a dc-CA model with a trait covariate
mod2 <- dc_CA_vegan(formulaEnv = ~A1+Moist+Mag+Use+Manure,
                   formulaTraits = ~ SLA+Height+ LDMC+ Lifespan +Condition(Seedmass) ,
                   response = dune_trait_env$comm[, -1],  # must delete "Sites"
                   dataEnv =dune_trait_env$envir,
                   dataTraits =dune_trait_env$traits,
                   verbose = TRUE)

# for illustration: a dc-CA model with both environmental and trait covariates
mod3 <- dc_CA_vegan(formulaEnv = ~A1+Moist+Use+Manure+Condition(Mag),
                    formulaTraits = ~ SLA+Height+LDMC+Lifespan +Condition(Seedmass) ,
                    response = dune_trait_env$comm[, -1],  # must delete "Sites"
                    dataEnv =dune_trait_env$envir,
                    dataTraits =dune_trait_env$traits,
                    verbose = TRUE)

# for illustration: same model but using mod2 for speed, as the trait model and data did not change
mod3B <- dc_CA_vegan(formulaEnv = ~A1+Moist+Use+Manure+Condition(Mag),
                    dataEnv =dune_trait_env$envir,
                    dc_CA_vegan_object = mod2,
                    verbose = TRUE)
all.equal(mod3,mod3B) # TRUE

mod_scores <- scores(mod3, display = "all", tidy = TRUE)
names(mod_scores)
levels(mod_scores$score)

mod_scores <- scores(mod3, display = "all")
str(mod_scores)


# All statistics and scores have been checked against the results with "focus on Case distances" (=Sites)
# in Canoco 5.15 (ter Braak & Smilauer, 1918).

}
